Add test for [makefilename] and correct implementation

+++ src/MessageMakefilename.cpp
@@ -38,8 +38,8 @@ void MessageMakefilename::processMessage(int inletIndex, PdMessage *message) {
   if (inletIndex == 0) {
     switch (message->getType(0)) {
       case FLOAT: {
-        char str[snprintf(NULL, 0, format, message->getFloat(0))+1];
-        snprintf(str, sizeof(str), format, message->getFloat(0));
+        char str[snprintf(NULL, 0, format, (int) message->getFloat(0))+1];
+        snprintf(str, sizeof(str), format, (int) message->getFloat(0));
         PdMessage *outgoingMessage = PD_MESSAGE_ON_STACK(1);
         outgoingMessage->initWithTimestampAndSymbol(message->getTimestamp(), str);
         sendMessage(0, outgoingMessage);
